Migrated hash-similarity-functions to corresponding class
